Add initial RO-Crate import
The required import format is generated from the RO-Crate of the Dataset, later this import format can be imported after modifying the "importRoCrate" function. Until then, this import format can be used to update the metadata for a dataset by calling the update mechanism from the cmdline with the import format that can be copied from the logs.