Add generation of OntoPortal based controlled vocabs to AROMA
Display of the controlled vocab works, however there are some internal
DV related problems, which creates a ControlledVocabularyValue for
each saved value in DV and afterwards only allows setting these values from
AROMA. The DV fields should be a primitive text field and handled internally
like so, and only show the slect in the UI based on the fact that this is
an ontoportal based controlled vocab. If it was a true primitive DV
datasetfieldtype, no ControlledVocabularyValue would be added for it.