Add authentication using CEDAR api key
In AuthenticatedUserArp we can associate the CEDAR api key with an
authenticated user and this key can then be used just like the DV api token
to authorize a user. This way the CEDAR login and DV login can be associated
and a CEDAR user doesn't have to authorize with DV in any other way.