Add version control for RO-Crates - refs #6862
When a Dataset is published, its corresponding RO-Crate will now be saved as well, using the same version number as the Dataset. When downloading the RO-Crate, the one from the latest published Dataset version will be retrieved, unless it doesn't exist, in which case the RO-Crate for the DRAFT version of the Dataset will be downloaded.