Add mdbsBasedConformsToIdsOfDataset and minimalDescriboProfileForDataset
We expect that for every entity, including the root dataset conformsTo
contains references to all the schemas applicable to the entity. Therefore
AROMA now uses minimalDescriboProfileForDataset to get a basic profile,
which is then improved with other schemas based on the selected entity's
conformsTo. The root dataset is however special, because it has some
schemas, which cannot be removed, because these are tied to the dataset's
dataverse's metadatablocks. mdbsBasedConformsToIdsOfDataset returns this
list and so AROMA can make sure that these schemas could not be removed
from the root datasets conformsTo list, while others can still be added
and removed.