Add RoCrateUploadServiceBean.addUploadedFileMetadata skeleton
This method should add file/sub-dataset metadata to a DV generated
RoCrate. This is necessary when an RO-Crate zip is uploaded and the first
time DV generates the ro-crate-metadata.json. This json is generated
based on only the root dataset and the generator is not aware of the
file metadata found in the uploaded ro-crate-metadata.json.
addUploadedFileMetadata will take care of adjusting the new metadata json
with the originally uploaded one.