Update RO-CRATE import from .zip files
Before the RO-CRATE is actually processed a mapping is created between the file entities from the RO-CRATE and their datasetFile representation for precise matching later when the files are being processed. Upon saving the uploaded RO-CRATE, first the originally uploaded RO-CRATE will be saved, then the modifications (if there were any) will be applied to the original RO-CRATE as if it were an update.