Fix NPE when publishing a dataset changed in the meantime
When a Dataset is published from UI, it uses a version of the dataset stored
in the session (in DatasetPage), while the underlying data in the db have
already changed. This can result in databse exception. To avoid this we
reload the dataset before publishing to make sure we work with its latest
state.