Add auto reloading dataset page if dataset changed in AROMA
It uses the CrossWindowMessenger class to detect changes as reported by
AROMA using this same class. Once change is detected the dataset page
(dataset.xhtml) is reloaded when the user switches back to it from AROMA.

Also added a feature to be able to target a specifc tab inside dataset.xhtml
by providing the name of the tab in the hash part of the URL, eg.

dataset.xhtml?...#metadataMapTab

will automatically switch to the Metadata tab.