Make MDB-CEDAR synchronization idempotent
/api/admin/arp/initialSetup has been deprecated and replaced with
/api/admin/arp/syncMdbsWithCedar

Admin.arpSyncMdbsWithCedar() creates CEDAR templates from MDB-s, uploads
them to CEDAR and syncs them back to Dataverse ensuring that both are in
sync. syncMdbsWithCedar is idempotent, which mean it can be called any number
of times and it will generate the same results, ie. templates and elements
with always the same identifiers.

Note: there are some values in CEDAR templates, which are not represented
in MDB-s, like display values for compound types. If you have such settings
in a CEDAR template and call syncMdbsWithCedar, these will be overriden
(removed).

Also made some adjustment to convertMdbToTsv, exportMdbToCedar, and
exportTsvToCedar so that they expect an MDB name instead of an id, which
makes it more straightforward to use in any installation where the MDB names
are ficed, while ID-s may differ.