Add arp.aroma.auth setting
arp.aroma.auth can have the value of either "apiKey" or "JSESSIONID". In case
of "apiKey" when generating an AROMA url we append an apiKey query paramater
with the api key of the currently logged in user, this way passing AROMA the
key to use to contact Dataverse. If the value is "JSESSIONID", then no apiKey
query paramater is generated and AROMA is assumed to be in the same domain as
DV and so the JSESSIONID is available to using. Using its value and the newly
added /api/arp/apiKey can grab the api key for DV communication.