Update ArpCedarIT to work with the latest updates in the code
To run the tests, you have to also run DV and CEDAR and provide the required values in the cedarParameters.json file.

The following properties are required:
"folderId" - the parent folder ID in CEDAR where the test results will be uploaded
"apiKey" - CEDAR API key of a user with write access to the parent folder
"cedarDomain" - CEDAR domain
"cedarUUID" - CEDAR UUID of a user with write access to the parent folder

The parent folder will be removed from CEDAR when the test is finished, stop the tests in debug mode if you want to check the folder contents.

TODO: Update the test schemas with their latest version from DV