Add support for CEDAR "checkbox" type in AROMA
Fix for https://redmine.sztaki.hu/issues/11903

CEDAR checkbox type is now mapped to CEDAR Select type. This is not ideal,
because in CEDAR this a multiselect list where the possible values are all
shown and one can select the values via checkboxes. As Describo deosn't have
this interface we can only map to the dropdown Select widget