Fix solr schema update and cache external cvv values - refs #25477
Upon CEDAR Schema import the solr schema was updated before all the imported data was updated in the db, this resulted in missing properties from the solr schema.
Checking external cvv-s during CEDAR Schema validation too, this way the data is checked against the correct values. (It was checked against the datasetfieldtype's cvv-s, that in case of external cvv-s meant only the values that have already been saved in dataverse for the given dsft.)