Fix rocrate validation - #25536
ro-crate-java 1.0.1 library uses a JSON Schema for RO-Crate validation that
is too restrictive and doesn't conform to RO-Crate specification. It only
allowed arrays with object or strings, but not integers. Newer version of
the lib (2.0.3) fixes it, ut at this time we didn't want to upgrade to it
because we have overriden files from that lib that needs to be adjusted.
So instead I just took the update JSON Schema files and overriden the
original ones with these new ones. The ultimate fix will be an upgrade
to a newer ro-crate-java version.

See also: #25536