ro-crate-java upgrade - refs #25630
Instead of using our custom RoCrateReader with a StringReader, ReadStringStrategy is used to process the RO-Crates in String format.

Still, a custom version of the PreviewGenerator is required.
For some reason the Jackson lib is not properly updated in the ro-crate-java lib that causes errors, so a custom version of AbstractEntity is required until it is fixed, that uses Jackson 2.19 from the DV dependencies.